creates the updates board api call
I updated the ajax calls to handle the update the board on the server
everytime that you click on a div. I added classes to the divs so that
it was easier to select them all. I was able to pass in the id the html
and the gameOn boolean into the update board as the index value and game
off on the server.