gets the new game reset button working
I was able to add to the onNewGame function in the game-events file
to re-set the board to be able to play multiple games. I had to add to
the game logic to get this working.. so i added some if statements to
check if there was anything in the html before you let someone put
a mark there. I did this instead of turning off my click handlers